requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-auth-ldap==4.5.0
django-debug-toolbar==4.2.0
django-widget-tweaks==1.5.0
idna==3.4
Markdown==3.4.4
mysqlclient==2.2.0
Pillow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyotp==2.9.0
pypng==0.20220715.0
python-ldap==3.4.3
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
slack-sdk==3.22.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
